Add files via upload
Question 1: 
Object oriented programming is a type of programming paradigm which is abbreviated as OOP. This approach of programming removes the limitations of procedural oriented programming.OOP treats data as an important element and does not allow it to flow freely around the program. Here basic entities called objects are used as data-structures that consist of both data and procedures.

Question 2:
Simplicity: software objects model real world objects, so the complexity is reduced and the program structure is very clear;
Modularity: each object forms a separate entity whose internal workings are decoupled from other parts of the system;
Modifiability: it is easy to make minor changes in the data representation or the procedures in an OO program. Changes inside a class do not affect any other part of a program, since the only public interface that the external world has to a class is through the use of methods;
Extensibility: adding new features or responding to changing operating environments can be solved by introducing a few new objects and modifying some existing ones;
Maintainability: objects can be maintained separately, making locating and fixing problems easier

Question 3: 
A function is a piece of code that is called by name. It can be passed data to operate on (i.e. the parameters) and can optionally return data (the return value). All data that is passed to a function is explicitly passed. A method is a piece of code that is called by a name that is associated with an object.

Question 4:
Class:
A class is a collection of method and variables. It is a blueprint that defines the data and behavior of a type. Like Human Being, having body parts, performing various actions. A class can be defined using the class keyword and the class body enclosed by a pair of curly braces.

Object:
Each object is an instance of a particular class or subclass with the class's own methods or procedures and data variables. An object is what actually runs in the computer.

Attribute:
An Attribute is a named property of a class. It has a type. It describes the range of values that that property may hold.

Behavior:
The behavior of an object is defined by its methods, which are the functions and subroutines defined within the object class. Without class methods, a class would simply be a structure. A class that inherits another class can also override methods of the inherited class to enhance or change its behavior. If the new class changes a method's behavior, it overrides the method and doesn't call the base class's method.